refactor: flatten package structure for simpler imports
This commit restructures the repository by moving code from the nested
pkg/anytype/ directory to the root level. This change simplifies imports
for users of the library, allowing direct imports from the main module
path without requiring extra path segments.

Changes include:
- Move core interface definitions to root directory
- Preserve implementation directories (client, middleware, options)
- Relocate tests and mocks to top-level directories
- Remove unnecessary nesting that complicated import statements

This change maintains all functionality while providing a cleaner,
more idiomatic Go package structure.
