+++ client/member.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/epheo/anytype-go/pkg/anytype"
-	"github.com/epheo/anytype-go/pkg/anytype/options"
+	"github.com/epheo/anytype-go"
+	"github.com/epheo/anytype-go/options"
 )
 
 // MemberClientImpl implements the MemberClient interface
