+++ tests/mocks/members_lists.go
@@ -3,7 +3,7 @@ package mocks
 import (
 	"context"
 
-	"github.com/epheo/anytype-go/pkg/anytype"
+	"github.com/epheo/anytype-go"
 )
 
 // MockMembersService implements the anytype.MemberClient interface for testing
