+++ tests/mocks/objects.go
@@ -3,8 +3,8 @@ package mocks
 import (
 	"context"
 
-	"github.com/epheo/anytype-go/pkg/anytype"
-	"github.com/epheo/anytype-go/pkg/anytype/options"
+	"github.com/epheo/anytype-go"
+	"github.com/epheo/anytype-go/options"
 )
 
 // MockObjectsService implements the anytype.ObjectClient interface for testing
