+++ tests_api_coverage/update_api_definition.sh
